# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "alwism.com - API Documentation",
        "description": "This is the API Gateway for [www.alwism.com](https://www.alwism.com).  You can find out more about alwism at [https://www.alwism.com/sitemechanism](https://www.alwism.com/sitemechanism) or visit to github repository [github](http://url.cc/).",
        "contact": {
            "name": "Contact the Developer",
            "url": "https://www.alwism.com/contactme"
        },
        "version": "1.0.3"
    },
    "servers": [
        {
            "url": "api/v1"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Non Auth"
                ],
                "summary": "Login",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "email": {
                                                "type": "string"
                                            },
                                            "password": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "example@gmail.com",
                                    "password": "your-secret"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "422": {
                        "description": "error"
                    }
                }
            }
        },
        "/api/admin/logout": {
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Logout User",
                "operationId": "db8d26e4c798faabe0c89ea0adfc8c31",
                "responses": {
                    "200": {
                        "description": "succes: client's action was received, understood, accepted and responded."
                    },
                    "401": {
                        "description": "unauthorized: Server requires the client to authorize, use login and token"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/admin/user": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Logged User",
                "operationId": "e8db7ba26333379d865b8539ba3c420a",
                "responses": {
                    "200": {
                        "description": "succes: client's action was received, understood, accepted and responded."
                    },
                    "401": {
                        "description": "unauthorized: Server requires the client to authorize, use login and token"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/admin/categories": {
            "get": {
                "tags": [
                    "categories"
                ],
                "summary": "Get categories",
                "operationId": "8bd209fd1d32f30bb6d46372608cc958",
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/200"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/admin/category": {
            "post": {
                "tags": [
                    "categories"
                ],
                "summary": "Create a new category.",
                "operationId": "e91983d01da5dbb5dc51e30a24bac7eb",
                "requestBody": {
                    "$ref": "#/components/requestBodies/Category"
                },
                "responses": {
                    "200": {
                        "description": "succes: client's action was received, understood, accepted and responded."
                    },
                    "401": {
                        "description": "unauthorized: Server requires the client to authorize, use login and token"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/admin/category/{id}": {
            "get": {
                "tags": [
                    "categories"
                ],
                "summary": "Get category by id",
                "operationId": "9897a42d312c0c924c7cc1ee34b79c56",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the category that needs to be retrive",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/200"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "categories"
                ],
                "summary": "Update an existing category",
                "operationId": "d0b18842ba2888b0b12f06561f8c8534",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the category that needs to be update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int",
                            "minimum": 1
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/Category"
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/200"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "categories"
                ],
                "summary": "Deletes a category",
                "operationId": "0d6e5831236c65c6fc12deab36929b27",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/200"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Category": {
                "title": "Category",
                "description": "Category model",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Category Name",
                        "type": "string",
                        "format": "string",
                        "default": "Nginx"
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
            "200": {
                "description": "succes: client's action was received, understood, accepted and responded."
            },
            "401": {
                "description": "unauthorized: Server requires the client to authorize. Login with email and password to get the authentication token"
            },
...
        },
        "requestBodies": {
...
            "Category": {
                "description": "Category Request Body",
                "required": true,
                "content": {
                    "application/x-www-form-urlencoded": {
                        "schema": {
                            "$ref": "#/components/schemas/Category"
                        }
                    },
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Category"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}